moved lecture stuff to another .js file for reference
